requirements.txt
--index-url http://192.168.186.125:8081/repository/aliyun-pypi/simple

aio-pika==9.2.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
aioboto3==11.2.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
aiobotocore[boto3]==2.5.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
aiohttp==3.8.5 ; python_full_version >= "3.9.12" and python_version < "3.10"
aioitertools==0.11.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
aiormq==6.7.7 ; python_full_version >= "3.9.12" and python_version < "3.10"
aiosignal==1.3.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
alembic==1.11.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
anyio==3.7.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
anyio[trio]==3.7.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
astor==0.8.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
async-generator==1.10 ; python_full_version >= "3.9.12" and python_version < "3.10"
async-timeout==4.0.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
attrs==23.1.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
bcrypt==4.0.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
boto3==1.26.76 ; python_full_version >= "3.9.12" and python_version < "3.10"
botocore==1.29.76 ; python_full_version >= "3.9.12" and python_version < "3.10"
certifi==2023.7.22 ; python_full_version >= "3.9.12" and python_version < "3.10"
cffi==1.15.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
charset-normalizer==3.2.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
click==8.1.7 ; python_full_version >= "3.9.12" and python_version < "3.10"
colorama==0.4.6 ; python_full_version >= "3.9.12" and python_version < "3.10" and platform_system == "Windows"
conductor-python==1.0.71 ; python_full_version >= "3.9.12" and python_version < "3.10"
cryptography==41.0.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
deprecation==2.1.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
elasticsearch6==6.8.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
et-xmlfile==1.1.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
exceptiongroup==1.1.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
fastapi-login==1.9.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
fastapi==0.94.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
filetype==1.2.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
frozenlist==1.4.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
ghostscript==0.7 ; python_full_version >= "3.9.12" and python_version < "3.10"
greenlet==2.0.2 ; python_full_version >= "3.9.12" and python_version < "3.10" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
h11==0.14.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
httpcore==0.16.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
httpx==0.23.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
idna==3.4 ; python_full_version >= "3.9.12" and python_version < "3.10"
jmespath==1.0.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
lxml==4.9.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
mako==1.2.4 ; python_full_version >= "3.9.12" and python_version < "3.10"
markupsafe==2.1.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
minio==7.1.16 ; python_full_version >= "3.9.12" and python_version < "3.10"
multidict==6.0.4 ; python_full_version >= "3.9.12" and python_version < "3.10"
numpy==1.23.5 ; python_full_version >= "3.9.12" and python_version < "3.10"
opencv-python==4.8.0.76 ; python_full_version >= "3.9.12" and python_version < "3.10"
openpyxl==3.1.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
outcome==1.2.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
packaging==23.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
pamqp==3.2.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
pandas==1.5.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
passlib[bcrypt]==1.7.4 ; python_full_version >= "3.9.12" and python_version < "3.10"
pdf2image==1.16.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
pdfminer-six==20221105 ; python_full_version >= "3.9.12" and python_version < "3.10"
pdfplumber==0.7.6 ; python_full_version >= "3.9.12" and python_version < "3.10"
pikepdf==6.2.9 ; python_full_version >= "3.9.12" and python_version < "3.10"
pillow==10.0.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
prometheus-client==0.17.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
psycopg2-binary==2.9.7 ; python_full_version >= "3.9.12" and python_version < "3.10"
pycparser==2.21 ; python_full_version >= "3.9.12" and python_version < "3.10"
pydantic==1.10.12 ; python_full_version >= "3.9.12" and python_version < "3.10"
pydantic[dotenv]==1.10.12 ; python_full_version >= "3.9.12" and python_version < "3.10"
pyhumps==3.8.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
pyjwt==2.8.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
pymupdf==1.23.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
pymupdfb==1.23.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
python-dateutil==2.8.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
python-dotenv==1.0.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
python-multipart==0.0.6 ; python_full_version >= "3.9.12" and python_version < "3.10"
pytz==2023.3 ; python_full_version >= "3.9.12" and python_version < "3.10"
requests==2.31.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
rfc3986[idna2008]==1.5.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
s3transfer==0.6.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
setuptools==68.1.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
six==1.16.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
sniffio==1.3.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
sortedcontainers==2.4.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
sqlalchemy==2.0.20 ; python_full_version >= "3.9.12" and python_version < "3.10"
starlette==0.26.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
trio==0.21.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
typing-extensions==4.7.1 ; python_full_version >= "3.9.12" and python_version < "3.10"
urllib3==1.26.16 ; python_full_version >= "3.9.12" and python_version < "3.10"
uvicorn==0.19.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
wand==0.6.11 ; python_full_version >= "3.9.12" and python_version < "3.10"
wrapt==1.15.0 ; python_full_version >= "3.9.12" and python_version < "3.10"
yarl==1.9.2 ; python_full_version >= "3.9.12" and python_version < "3.10"
